+++ NEWS.md
@@ -1,8 +1,8 @@
 # connectapi (development version)
 
-- New `content_set_all_integrations()` function to set the OAuth integration
+- New `content_set_integrations()` function to set the OAuth integration
   associations for a content item. (#414)
-- New `integration()` function to retrieve details of a specific OAuth integration
+- New `get_integration()` function to retrieve details of a specific OAuth integration
   from a Connect server. (#431)
 
 # connectapi 0.8.0